requirements.txt
biopython==1.85
ete3==3.1.3
joblib==1.5.2
numpy==2.2.6
six==1.17.0
torch==2.0.1
matplotlib==3.8.0
